Tinggi-Badan-Ideal/requirement.txt
blinker==1.8.2
bootstraps==1.0.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==10.4.0
ttkbootstrap==1.12.2
Werkzeug==3.1.1
